+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.3] - 2021-07-31
+### Added
+- Send zabbix as smb user for gettings server shares
+- Manage space in SMB share names
+- Change separator from ";" to "," for more compatibility 
+
 ## [1.0.2] - 2020-03-09
 ### Added
 - Fix trigger not triggering Problems